pyproject.toml
[tool.poetry]
name = "google driver manager"
version = "0.1.0"
description = ""
authors = ["Jonas Lima <jonas.lima.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydrive = "^1.3.1"
python-magic = "^0.4.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
